# backend/GinSqlBlog.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GinSqlBlog",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/user/login": {
      "post": {
        "summary": "用户登录接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginReq"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginRsp"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/postModify": {
      "post": {
        "summary": "用户报文接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sqlReq"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sqlRsp"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "loginReq": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "loginRsp": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "sqlReq": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "appendText": {
            "type": "string"
          }
        },
        "required": [
          "mode",
          "token",
          "userName",
          "appendText"
        ]
      },
      "sqlRsp": {
        "type": "object",
        "properties": {
          "context": {
            "type": "string"
          }
        },
        "required": [
          "context"
        ]
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}